catalog: Move proto objects from stash to catalog
This commit moves all catalog specific protobuf objects from the stash
crate to the catalog crate. Additionally, it moves all the logic for
serializing those objects from Rust objects to Protobuf objects to the
catalog crate. Specifically, this involves moving a lot of
`impl RustType<_> for _`. Previously those were spread out across a
couple of crates, and now they are all in the catalog crate.

Also, this commit moves the old protobuf snapshots and migrations from
the stash crate to the catalog crate. This will help us generalize
catalog data migrations to work with the persist catalog.

Works towards resolving #22977
