+++ src/catalog/protos/objects_v41.proto
