+++ src/catalog/protos/objects_v42.proto
