+++ src/catalog/protos/objects_v43.proto
