+++ src/catalog/src/durable/upgrade/v41_to_v42.rs
