+++ src/catalog/src/durable/upgrade/v42_to_v43.rs
