+++ src/cluster-client/Cargo.toml
@@ -15,7 +15,6 @@ http = "0.2.8"
 itertools = "0.10.5"
 mz-ore = { path = "../ore", features = ["tracing_"] }
 mz-proto = { path = "../proto" }
-mz-stash-types = { path = "../stash-types" }
 once_cell = "1.16.0"
 prometheus = { version = "0.13.3", default-features = false }
 proptest = { version = "1.0.0", default-features = false, features = ["std"]}
