+++ src/cluster-client/src/client.rs
@@ -215,8 +215,6 @@ pub struct ClusterReplicaLocation {
 
 #[cfg(test)]
 mod tests {
-    use mz_proto::protobuf_roundtrip;
-    use proptest::prelude::ProptestConfig;
     use proptest::proptest;
 
     use super::*;
