+++ src/sql/Cargo.toml
@@ -43,7 +43,6 @@ mz-secrets = { path = "../secrets" }
 mz-sql-parser = { path = "../sql-parser" }
 mz-sql-lexer = { path = "../sql-lexer" }
 mz-ssh-util = { path = "../ssh-util" }
-mz-stash-types = { path = "../stash-types" }
 mz-storage-types = { path = "../storage-types" }
 mz-tracing = { path = "../tracing" }
 paste = "1.0"
