+++ src/sql/src/names.rs
@@ -18,12 +18,10 @@ use mz_controller_types::{ClusterId, ReplicaId};
 use mz_expr::LocalId;
 use mz_ore::cast::CastFrom;
 use mz_ore::str::StrExt;
-use mz_proto::{IntoRustIfSome, ProtoType};
 use mz_repr::role_id::RoleId;
 use mz_repr::ColumnName;
 use mz_repr::GlobalId;
 use mz_sql_parser::ast::{MutRecBlock, UnresolvedObjectName};
-use mz_stash_types::objects::{proto, RustType, TryFromProtoError};
 use once_cell::sync::Lazy;
 use proptest_derive::Arbitrary;
 use serde::{Deserialize, Serialize};
@@ -301,35 +299,6 @@ impl From<Option<DatabaseId>> for ResolvedDatabaseSpecifier {
     }
 }
 
-impl RustType<proto::ResolvedDatabaseSpecifier> for ResolvedDatabaseSpecifier {
-    fn into_proto(&self) -> proto::ResolvedDatabaseSpecifier {
-        let spec = match self {
-            ResolvedDatabaseSpecifier::Ambient => {
-                proto::resolved_database_specifier::Spec::Ambient(Default::default())
-            }
-            ResolvedDatabaseSpecifier::Id(database_id) => {
-                proto::resolved_database_specifier::Spec::Id(database_id.into_proto())
-            }
-        };
-        proto::ResolvedDatabaseSpecifier { spec: Some(spec) }
-    }
-
-    fn from_proto(proto: proto::ResolvedDatabaseSpecifier) -> Result<Self, TryFromProtoError> {
-        let spec = proto
-            .spec
-            .ok_or_else(|| TryFromProtoError::missing_field("ResolvedDatabaseSpecifier::spec"))?;
-        let spec = match spec {
-            proto::resolved_database_specifier::Spec::Ambient(_) => {
-                ResolvedDatabaseSpecifier::Ambient
-            }
-            proto::resolved_database_specifier::Spec::Id(database_id) => {
-                ResolvedDatabaseSpecifier::Id(database_id.into_rust()?)
-            }
-        };
-        Ok(spec)
-    }
-}
-
 /*
  * TODO(jkosh44) It's possible that in order to fix
  * https://github.com/MaterializeInc/materialize/issues/8805 we will need to assign temporary
@@ -408,33 +377,6 @@ impl From<SchemaSpecifier> for SchemaId {
             SchemaSpecifier::Temporary => SchemaId::User(SchemaSpecifier::TEMPORARY_SCHEMA_ID),
             SchemaSpecifier::Id(id) => id,
         }
-    }
-}
-
-impl RustType<proto::SchemaSpecifier> for SchemaSpecifier {
-    fn into_proto(&self) -> proto::SchemaSpecifier {
-        let spec = match self {
-            SchemaSpecifier::Temporary => {
-                proto::schema_specifier::Spec::Temporary(Default::default())
-            }
-            SchemaSpecifier::Id(schema_id) => {
-                proto::schema_specifier::Spec::Id(schema_id.into_proto())
-            }
-        };
-        proto::SchemaSpecifier { spec: Some(spec) }
-    }
-
-    fn from_proto(proto: proto::SchemaSpecifier) -> Result<Self, TryFromProtoError> {
-        let spec = proto
-            .spec
-            .ok_or_else(|| TryFromProtoError::missing_field("SchemaSpecifier::spec"))?;
-        let spec = match spec {
-            proto::schema_specifier::Spec::Temporary(_) => SchemaSpecifier::Temporary,
-            proto::schema_specifier::Spec::Id(schema_id) => {
-                SchemaSpecifier::Id(schema_id.into_rust()?)
-            }
-        };
-        Ok(spec)
     }
 }
 
@@ -899,28 +841,6 @@ impl FromStr for SchemaId {
     }
 }
 
-impl RustType<proto::SchemaId> for SchemaId {
-    fn into_proto(&self) -> proto::SchemaId {
-        let value = match self {
-            SchemaId::User(id) => proto::schema_id::Value::User(*id),
-            SchemaId::System(id) => proto::schema_id::Value::System(*id),
-        };
-
-        proto::SchemaId { value: Some(value) }
-    }
-
-    fn from_proto(proto: proto::SchemaId) -> Result<Self, TryFromProtoError> {
-        let value = proto
-            .value
-            .ok_or_else(|| TryFromProtoError::missing_field("SchemaId::value"))?;
-        let id = match value {
-            proto::schema_id::Value::User(id) => SchemaId::User(id),
-            proto::schema_id::Value::System(id) => SchemaId::System(id),
-        };
-        Ok(id)
-    }
-}
-
 /// The identifier for a database.
 #[derive(
     Clone, Copy, Debug, Eq, PartialEq, Ord, PartialOrd, Hash, Serialize, Deserialize, Arbitrary,
@@ -967,25 +887,6 @@ impl FromStr for DatabaseId {
                 "couldn't parse DatabaseId {}",
                 s
             ))),
-        }
-    }
-}
-
-impl RustType<proto::DatabaseId> for DatabaseId {
-    fn into_proto(&self) -> proto::DatabaseId {
-        let value = match self {
-            DatabaseId::User(id) => proto::database_id::Value::User(*id),
-            DatabaseId::System(id) => proto::database_id::Value::System(*id),
-        };
-
-        proto::DatabaseId { value: Some(value) }
-    }
-
-    fn from_proto(proto: proto::DatabaseId) -> Result<Self, TryFromProtoError> {
-        match proto.value {
-            Some(proto::database_id::Value::User(id)) => Ok(DatabaseId::User(id)),
-            Some(proto::database_id::Value::System(id)) => Ok(DatabaseId::System(id)),
-            None => Err(TryFromProtoError::missing_field("DatabaseId::value")),
         }
     }
 }
@@ -1268,128 +1169,6 @@ pub enum CommentObjectId {
     Schema((ResolvedDatabaseSpecifier, SchemaSpecifier)),
     Cluster(ClusterId),
     ClusterReplica((ClusterId, ReplicaId)),
-}
-
-impl RustType<proto::comment_key::Object> for CommentObjectId {
-    fn into_proto(&self) -> proto::comment_key::Object {
-        match self {
-            CommentObjectId::Table(global_id) => {
-                proto::comment_key::Object::Table(global_id.into_proto())
-            }
-            CommentObjectId::View(global_id) => {
-                proto::comment_key::Object::View(global_id.into_proto())
-            }
-            CommentObjectId::MaterializedView(global_id) => {
-                proto::comment_key::Object::MaterializedView(global_id.into_proto())
-            }
-            CommentObjectId::Source(global_id) => {
-                proto::comment_key::Object::Source(global_id.into_proto())
-            }
-            CommentObjectId::Sink(global_id) => {
-                proto::comment_key::Object::Sink(global_id.into_proto())
-            }
-            CommentObjectId::Index(global_id) => {
-                proto::comment_key::Object::Index(global_id.into_proto())
-            }
-            CommentObjectId::Func(global_id) => {
-                proto::comment_key::Object::Func(global_id.into_proto())
-            }
-            CommentObjectId::Connection(global_id) => {
-                proto::comment_key::Object::Connection(global_id.into_proto())
-            }
-            CommentObjectId::Type(global_id) => {
-                proto::comment_key::Object::Type(global_id.into_proto())
-            }
-            CommentObjectId::Secret(global_id) => {
-                proto::comment_key::Object::Secret(global_id.into_proto())
-            }
-            CommentObjectId::Role(role_id) => {
-                proto::comment_key::Object::Role(role_id.into_proto())
-            }
-            CommentObjectId::Database(database_id) => {
-                proto::comment_key::Object::Database(database_id.into_proto())
-            }
-            CommentObjectId::Schema((database, schema)) => {
-                proto::comment_key::Object::Schema(proto::ResolvedSchema {
-                    database: Some(database.into_proto()),
-                    schema: Some(schema.into_proto()),
-                })
-            }
-            CommentObjectId::Cluster(cluster_id) => {
-                proto::comment_key::Object::Cluster(cluster_id.into_proto())
-            }
-            CommentObjectId::ClusterReplica((cluster_id, replica_id)) => {
-                let cluster_replica_id = proto::ClusterReplicaId {
-                    cluster_id: Some(cluster_id.into_proto()),
-                    replica_id: Some(replica_id.into_proto()),
-                };
-                proto::comment_key::Object::ClusterReplica(cluster_replica_id)
-            }
-        }
-    }
-
-    fn from_proto(proto: proto::comment_key::Object) -> Result<Self, TryFromProtoError> {
-        let id = match proto {
-            proto::comment_key::Object::Table(global_id) => {
-                CommentObjectId::Table(global_id.into_rust()?)
-            }
-            proto::comment_key::Object::View(global_id) => {
-                CommentObjectId::View(global_id.into_rust()?)
-            }
-            proto::comment_key::Object::MaterializedView(global_id) => {
-                CommentObjectId::MaterializedView(global_id.into_rust()?)
-            }
-            proto::comment_key::Object::Source(global_id) => {
-                CommentObjectId::Source(global_id.into_rust()?)
-            }
-            proto::comment_key::Object::Sink(global_id) => {
-                CommentObjectId::Sink(global_id.into_rust()?)
-            }
-            proto::comment_key::Object::Index(global_id) => {
-                CommentObjectId::Index(global_id.into_rust()?)
-            }
-            proto::comment_key::Object::Func(global_id) => {
-                CommentObjectId::Func(global_id.into_rust()?)
-            }
-            proto::comment_key::Object::Connection(global_id) => {
-                CommentObjectId::Connection(global_id.into_rust()?)
-            }
-            proto::comment_key::Object::Type(global_id) => {
-                CommentObjectId::Type(global_id.into_rust()?)
-            }
-            proto::comment_key::Object::Secret(global_id) => {
-                CommentObjectId::Secret(global_id.into_rust()?)
-            }
-            proto::comment_key::Object::Role(role_id) => {
-                CommentObjectId::Role(role_id.into_rust()?)
-            }
-            proto::comment_key::Object::Database(database_id) => {
-                CommentObjectId::Database(database_id.into_rust()?)
-            }
-            proto::comment_key::Object::Schema(resolved_schema) => {
-                let database = resolved_schema
-                    .database
-                    .into_rust_if_some("ResolvedSchema::database")?;
-                let schema = resolved_schema
-                    .schema
-                    .into_rust_if_some("ResolvedSchema::schema")?;
-                CommentObjectId::Schema((database, schema))
-            }
-            proto::comment_key::Object::Cluster(cluster_id) => {
-                CommentObjectId::Cluster(cluster_id.into_rust()?)
-            }
-            proto::comment_key::Object::ClusterReplica(cluster_replica_id) => {
-                let cluster_id = cluster_replica_id
-                    .cluster_id
-                    .into_rust_if_some("ClusterReplicaId::cluster_id")?;
-                let replica_id = cluster_replica_id
-                    .replica_id
-                    .into_rust_if_some("ClusterReplicaId::replica_id")?;
-                CommentObjectId::ClusterReplica((cluster_id, replica_id))
-            }
-        };
-        Ok(id)
-    }
 }
 
 #[derive(Debug)]
