+++ src/stash-types/Cargo.toml
@@ -17,20 +17,11 @@ proptest = { version = "1.0.0", default-features = false, features = ["std"] }
 proptest-derive = { version = "0.3.0", features = ["boxed_union"] }
 prost = { version = "0.11.9", features = ["no-recursion-limit"] }
 serde = "1.0.152"
-timely = { version = "0.12.0", default-features = false }
 tokio-postgres = { version = "0.7.8", features = ["with-serde_json-1"] }
 workspace-hack = { version = "0.0.0", path = "../workspace-hack" }
 
 [dev-dependencies]
 similar-asserts = "1.4"
-
-[build-dependencies]
-anyhow = "1.0.66"
-md-5 = "0.10.5"
-protobuf-src = "1.1.0"
-prost-build = "0.11.9"
-serde = "1.0.152"
-serde_json = "1.0.89"
 
 [package.metadata.cargo-udeps.ignore]
 normal = ["workspace-hack"]
